pyproject.toml
[project]
name = "memc_load"
version = "0.1.0"
description = ""
authors = [
    {name = "aleksei",email = "bscadet@yandex.ru"}
]
readme = "README.md"

packages = [{ include = "memc_load", from = "src" }]

requires-python = ">=3.13"
dependencies = [
    "protobuf (>=6.31.1,<7.0.0)",
    "python-memcached (>=1.62,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.16.0"
grpcio-tools = "^1.72.1"

